If compiler wrapper doesn't give MPI name try to use -show
As a fallback option, use the -show argument to MPI to try and find
the MPI family.

Default is "Unknown"

Currently this is only important for gfortran > 10 and dev builds
with only OpenMPI supplying the necessary modules to turn up the
warning/error flags fully. If the MPI family is MPICH or Unknown
then less strict flags are used.